# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Autores",
        "description": "API para gerenciar autores",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/book-borrowings/store": {
            "post": {
                "tags": [
                    "BookBorrowing"
                ],
                "summary": "Empresat um livro",
                "operationId": "74bff9e4c3a846c0dc9034a1f951298f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "book_id",
                                    "user_id",
                                    "from",
                                    "to"
                                ],
                                "properties": {
                                    "book_id": {
                                        "description": "1",
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "description": "1",
                                        "type": "string"
                                    },
                                    "from": {
                                        "description": "2024-12-01",
                                        "type": "string"
                                    },
                                    "to": {
                                        "description": "2024-12-01",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "livro emprestado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Record saved successfully!"
                                        },
                                        "data": {
                                            "properties": {
                                                "book_id": {
                                                    "description": "1",
                                                    "type": "string"
                                                },
                                                "user_id": {
                                                    "description": "1",
                                                    "type": "string"
                                                },
                                                "from": {
                                                    "description": "2024-12-01",
                                                    "type": "string"
                                                },
                                                "to": {
                                                    "description": "2024-12-01",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao salvar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error trying to save."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}